Refactor packages, add model classes, create callback interface
