+++ .idea/gradle.xml
@@ -8,7 +8,7 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/identsdk" />
+            <option value="$PROJECT_DIR$/identSDK" />
             <option value="$PROJECT_DIR$/sample" />
           </set>
         </option>
