+++ identSDK/src/main/java/com/twigbit/identsdk/dropinui/ErrorFragment.kt
@@ -1,4 +1,4 @@
-package com.twigbit.identsdk.ui
+package com.twigbit.identsdk.dropinui
 
 
 import android.os.Bundle
