+++ identSDK/src/main/java/com/twigbit/identsdk/dropinui/LoaderFragment.kt
@@ -1,4 +1,4 @@
-package com.twigbit.identsdk.ui
+package com.twigbit.identsdk.dropinui
 
 
 import android.os.Bundle
@@ -18,14 +18,14 @@ private const val ARG_PARAM2 = "param2"
  * A simple [Fragment] subclass.
  *
  */
-class SuccessFragment : Fragment() {
+class LoaderFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_success, container, false)
+        return inflater.inflate(R.layout.fragment_loader, container, false)
     }
