+++ identSDK/src/main/java/com/twigbit/identsdk/dropinui/SuccessFragment.kt
@@ -1,4 +1,4 @@
-package com.twigbit.identsdk.ui
+package com.twigbit.identsdk.dropinui
 
 
 import android.os.Bundle
@@ -9,18 +9,24 @@ import android.view.ViewGroup
 
 import com.twigbit.identsdk.R
 
+// TODO: Rename parameter arguments, choose names that match
+// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+private const val ARG_PARAM1 = "param1"
+private const val ARG_PARAM2 = "param2"
+
 /**
  * A simple [Fragment] subclass.
  *
  */
-class IntroFragment : Fragment() {
+class SuccessFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_intro, container, false)
+        return inflater.inflate(R.layout.fragment_success, container, false)
     }
+
 
 }
