+++ identSDK/src/main/java/com/twigbit/identsdk/model/IdentificationCard.kt
@@ -1,0 +1,5 @@
+package com.twigbit.identsdk.model
+
+class IdentificationCard{
+
+}
