+++ identSDK/src/main/java/com/twigbit/identsdk/model/IdentificationError.kt
@@ -1,0 +1,5 @@
+package com.twigbit.identsdk.model
+
+class IdentificationError{
+
+}
