+++ identsdk/src/main/res/layout/activity_dropin_identification.xml
@@ -5,6 +5,6 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.DropinIdentificationActivity">
+        tools:context=".dropinui.DropinIdentificationActivity">
 
 </android.support.constraint.ConstraintLayout>
