+++ identsdk/src/main/res/layout/fragment_loader.xml
@@ -3,7 +3,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context=".ui.LoaderFragment">
+             tools:context=".dropinui.LoaderFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
