+++ identsdk/src/main/res/layout/fragment_success.xml
@@ -3,7 +3,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context=".ui.SuccessFragment">
+             tools:context=".dropinui.SuccessFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
