+++ sample/src/main/java/com/twigbit/identsdk/sample/MainActivity.kt
@@ -1,9 +1,9 @@
 package com.twigbit.identsdk.sample
 
 import android.os.Bundle
-import com.twigbit.identsdk.IdentificationActivity
-import com.twigbit.identsdk.Message
-import com.twigbit.identsdk.DropInRequest
+import com.twigbit.identsdk.util.IdentificationActivity
+import com.twigbit.identsdk.util.Message
+import com.twigbit.identsdk.dropinui.DropInRequest
 import android.app.Activity
 import android.content.Intent
